requirements.txt
bottle==0.12.19
numpy==1.19.5
keras==2.4.3
Pillow==8.3.1
gunicorn
